requirements.txt
# Computer Use Assistant (CUA) for HPS Testing
# Dependencies matching Azure CUA sample: https://github.com/Azure-Samples/computer-use-model

# Core CUA dependencies (exact match with Azure sample)
openai>=1.68.2
pyautogui>=0.9.54
Pillow>=11.1.0

# Additional dependencies for enhanced functionality
click>=8.0.0
structlog>=23.1.0
pyyaml>=6.0
aiohttp>=3.8.0

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
